+++ R/edit_matches.R
@@ -77,37 +77,38 @@ edit_matches <- function(concepts, attributes = NULL, source = NULL,
   # determine those concepts, that are not yet defined in the ontology
   if(!is.null(prevMatches)){
 
-    temp <- temp %>%
-      rename(prev_broader_match = has_broader_match,
-             prev_close_match = has_close_match,
-             prev_exact_match = has_exact_match,
-             prev_narrower_match = has_narrower_match) %>%
-      full_join(prevMatches, by = c("label", "class", "id", "has_broader", "description")) %>%
-      unite(col = "has_broader_match", prev_broader_match, has_broader_match, sep = " | ", na.rm = TRUE) %>%
-      unite(col = "has_close_match", prev_close_match, has_close_match, sep = " | ", na.rm = TRUE) %>%
-      unite(col = "has_exact_match", prev_exact_match, has_exact_match, sep = " | ", na.rm = TRUE) %>%
-      unite(col = "has_narrower_match", prev_narrower_match, has_narrower_match, sep = " | ", na.rm = TRUE) %>%
-      na_if(y = "")
+    # temp <- temp %>%
+    #   rename(prev_broader_match = has_broader_match,
+    #          prev_close_match = has_close_match,
+    #          prev_exact_match = has_exact_match,
+    #          prev_narrower_match = has_narrower_match) %>%
+    #   full_join(prevMatches, by = c("label", "class", "id", "has_broader", "description")) %>%
+    #   unite(col = "has_broader_match", prev_broader_match, has_broader_match, sep = " | ", na.rm = TRUE) %>%
+    #   unite(col = "has_close_match", prev_close_match, has_close_match, sep = " | ", na.rm = TRUE) %>%
+    #   unite(col = "has_exact_match", prev_exact_match, has_exact_match, sep = " | ", na.rm = TRUE) %>%
+    #   unite(col = "has_narrower_match", prev_narrower_match, has_narrower_match, sep = " | ", na.rm = TRUE) %>%
+    #   na_if(y = "")
 
-    # temp <- prevMatches %>%
-    #   filter(class %in% filterClasses) %>%
-    #   rename(harmLab = label) %>%
-    #   pivot_longer(cols = c(has_broader_match, has_close_match, has_exact_match, has_narrower_match),
-    #                names_to = "match", values_to = "label") %>%
-    #   separate_rows(label, sep = " \\| ") %>%
-    #   full_join(concepts, by = "label") %>%
-    #   mutate(harmLab = if_else(is.na(match), label, harmLab),
-    #          label = if_else(is.na(match), NA_character_, label),
-    #          match = if_else(is.na(match), "sort_in", match)) %>%
-    #   pivot_wider(id_cols = c(harmLab, class, id, has_broader, description), names_from = match,
-    #               values_from = label, values_fn = ~paste0(.x, collapse = " | ")) %>%
-    #   na_if(y = "NA") %>%
-    #   rename(label = harmLab)
-    #
-    # if("sort_in" %in% colnames(temp)){
-    #   temp <- temp %>%
-    #     select(-sort_in)
-    # }
+    temp <- prevMatches %>%
+      filter(class %in% filterClasses) %>%
+      rename(harmLab = label) %>%
+      pivot_longer(cols = c(has_broader_match, has_close_match, has_exact_match, has_narrower_match),
+                   names_to = "match", values_to = "label") %>%
+      separate_rows(label, sep = " \\| ") %>%
+      # full_join(concepts, by = "label") %>%
+      full_join(temp, by = c("label", "class", "id", "has_broader", "description")) %>%
+      mutate(harmLab = if_else(is.na(match), label, harmLab),
+             match = if_else(!is.na(has_close_match), "has_close_match", if_else(is.na(match), "sort_in", match)),
+             label = if_else(is.na(match), NA_character_, label)) %>%
+      pivot_wider(id_cols = c(harmLab, class, id, has_broader, description), names_from = match,
+                  values_from = label, values_fn = ~paste0(.x, collapse = " | ")) %>%
+      na_if(y = "NA") %>%
+      rename(label = harmLab)
+
+    if("sort_in" %in% colnames(temp)){
+      temp <- temp %>%
+        select(-sort_in)
+    }
 
   }